State now correctly uses 255 instead of -1

+++ backend/client/src/main/java/be/xhibit/teletask/client/message/LogMessage.java
@@ -15,7 +15,7 @@ public class LogMessage extends MessageSupport {
 
     @Override
     protected byte[] getPayload() {
-        return new byte[]{this.state.getCode()};
+        return new byte[]{(byte) this.state.getCode()};
     }
 
     @Override
